% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dating.R
\name{beast_phydyn}
\alias{beast_phydyn}
\title{Convert tips of a phylogentic tree suitable to use with BEAST and phydyn}
\usage{
beast_phydyn(m.df1, m.cgr.df2, df3, df4, all_data)
}
\arguments{
\item{m.df1}{Dataframe object containig information for the Senegal HIV
samples in which metadata associated to it is missing.}

\item{m.cgr.df2}{Dataframe object containing information for the CGR sequences
that metadata associated to it is missing.}

\item{df3}{Dataframe object for all Senegal sequences}

\item{df4}{Dataframe object for all CGR sequences}

\item{all_data}{Dataframe containing all traits associated to each tip in the
phylogenetic tree}
}
\value{
a dataframe with additional columns representing the tips as
    expected by phydyn in BEAST
}
\description{
Function that makes some manipulations on dataframes to return a tip name
that is suitable to use with phydyn in BEAST. This function was written
taking into consideration this specific project and might not work as it is
for your own data.
}
\examples{
TO DO
}
