% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{bindTree_for_phydynBeast}
\alias{bindTree_for_phydynBeast}
\title{Merged phylogenetic trees for the 3 HIV-1 subtypes (B, C, and 02_AG)}
\description{
This is a phylogenetic tree to be used with phydyn in BEAST.
}
\section{bindTree_CGR_GTR+Gp12+3_droppedTip_phydynBeast.tre}{


This phylogenetic tree is in the Newick format, and was generated to be used
with phydyn in BEAST2. Tip names is such that allows BEAST to get the age of
the tips and phydyn to get the trait in the model. Trait in our example is
gpm, gpf, msm, and src. For more details see the vignette.
}

