% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{births_mq}
\alias{births_mq}
\title{Calculate median and quantiles when using births}
\usage{
births_mq(birth_list, times)
}
\arguments{
\item{birth_list}{list for births}

\item{times}{time points associated to the birth list}
}
\value{
data.frame
}
\description{
Calculate median and quantiles when using births
}
\examples{
# TO DO
}
