% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{births_newCases}
\alias{births_newCases}
\title{Organizes the solved births and calculate new cases}
\usage{
births_newCases(birth_list, times)
}
\arguments{
\item{birth_list}{list for births}

\item{times}{time points associated to the birth list}
}
\value{
data.frame
}
\description{
Organizes the solved births and calculate new cases
}
\examples{
# TO DO
}
