% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{births_pafs}
\alias{births_pafs}
\title{Organizes the solved birth element and calculate pafs}
\usage{
births_pafs(birth_list, times)
}
\arguments{
\item{birth_list}{list for births}

\item{times}{time points associated to the birth list}
}
\value{
data.frame
}
\description{
Organizes the solved birth element and calculate pafs
}
\examples{
# TO DO
}
