% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{calculate_newCases}
\alias{calculate_newCases}
\title{Calculate new cases}
\usage{
calculate_newCases(f)
}
\arguments{
\item{f}{list of births}
}
\value{
new cases
}
\description{
Calculate new cases
}
\examples{
# TO DO
}
