% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_posterior.R
\name{deme2deme}
\alias{deme2deme}
\title{Gets the proportion of infections in deme1 atributable to deme2}
\usage{
deme2deme(births.p, times, r, c)
}
\arguments{
\item{births.p}{list for birth elements (resulted from calculating the
demographic model)}

\item{times}{vector of time points associated to the birth lists}

\item{r}{row number for deme2}

\item{c}{column number for deme1}
}
\value{
dataframe of median, and lower and upper bounds for the proportion of
   infections in one deme attributable to another deme
}
\description{
Gets the proportion of infections in one deme attributable to another deme,
and calculates the median, lower and upper bounds.
}
\details{
row and column numbers will determine which proportion of infections
   in one deme is attributable to onother deme. For example, for our analysis
   for the Senegal data, \eqn{r = 3} and \eqn{c = 2} represent the proportion
   of infections in gpf attributable to msm
}
\examples{
#TO DO
}
