% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_posterior.R
\name{df_newCases}
\alias{df_newCases}
\title{Create dataframe for new cases of HIV}
\usage{
df_newCases(newCases.p, newCases.map, times)
}
\arguments{
\item{newCases.p}{list for birth elements (resulted from calculating the
demographic model)}

\item{newCases.map}{list for birth element for MAP (Maximum a posteriori)}

\item{times}{vector of time points associated to the birth list}
}
\value{
dataframe of median, and upper and lower bounds for the posterior,
   and the same for MAP
}
\description{
Creates a dataframe for new cases of HIV. This
dataframe have information for median and lower and upper bounds for new HIV
cases by group (gpm, gpf and msm).
}
\examples{
#TO DO
}
