% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_posterior.R
\name{df_pafs}
\alias{df_pafs}
\title{Create dataframe for PAF}
\usage{
df_pafs(births.p, births.map, times)
}
\arguments{
\item{births.p}{list for birth elements (resulted from calculating the
demographic model)}

\item{births.map}{list for birth elements for MAP (Maximum a posteriori)}

\item{times}{vector of time points associated to the birth lists}
}
\value{
dataframe of median and upper, lower bounds for the posterior, and
   PAF for MAP
}
\description{
Creates a dataframe for population attributable fraction (PAF). This
dataframe have information for median and lower and upper bounds for PAF
by group (gpm, gpf and msm).
}
\examples{
#TO DO
}
