% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_posterior.R
\name{df_r0}
\alias{df_r0}
\title{Create dataframe for basic reproduction number}
\usage{
df_r0(run, burnin, par_names, times, T0, T1, GAMMA)
}
\arguments{
\item{run}{object of class "mcmcSamplerList" and "bayesianOutput"}

\item{burnin}{intenger representing number of samples to remove as burnin}

\item{par_names}{vector containinig the parameter names}

\item{times}{vector of time points}

\item{T0}{initial time for simulations}

\item{T1}{end time for simulations}

\item{GAMMA}{double representing the stage of infection}
}
\value{
dataframe of median, and upper and lower bounds for the posterior,
   and MAP
}
\description{
Create dataframe for basic reproduction number (R0)
}
\details{
Basic Reproduction Number (R0) is calculated based on the equation
   \eqn{R0 = \beta / \gamma}. In our model \eqn{\gamma = 0.1} and \eqn{\beta}
    is represented by each of the linear function parameters
    (ex: gpsp0, gpsp1, etc). \eqn{\beta} is the number of transmissions per
    infected individuals.
}
\examples{
#TO DO
}
