% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_posterior.R
\name{get_posterior}
\alias{get_posterior}
\title{Get posterior distributions of MCMC runs}
\usage{
get_posterior(run, burnin, rep, par_names)
}
\arguments{
\item{run}{object of class "mcmcSamplerList" and "bayesianOutput"}

\item{burnin}{intenger representing number of samples to remove as burnin}

\item{rep}{string representing replicate of MCMC run}

\item{par_names}{vector containinig the parameter names}
}
\value{
a dataframe containing posterior distributions for each paramater
   in the model, with an additional column indicating the replicate number.
}
\description{
Gets posterior distributions of the Markov chain Monte Carlo (MCMC) runs
}
\examples{
# TO DO
}
