% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{gpspline}
\alias{gpspline}
\title{Calculate the shape of the linear function for the general population}
\usage{
gpspline(t, T0, T1, parms)
}
\arguments{
\item{t}{a time point}

\item{T0}{initial time for simulations}

\item{T1}{end time for simulations}

\item{parms}{list of 4 parameters of the linear functions}
}
\value{
list of points which linearly interpolate with the given data points
}
\description{
Calculate the shape of the linear function for the general population
}
\examples{
#TO DO
}
