% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{mq_prop}
\alias{mq_prop}
\title{Calculates the proportion of each deme}
\usage{
mq_prop(sizes_list, times)
}
\arguments{
\item{sizes_list}{list containg the reorganized element size (effective number
of infections).}

\item{times}{vector containig the time points used in the simulations}
}
\value{
dataframe for median and quantiles for each deme (gpm, gpf and msm)
   in proportions
}
\description{
Calculates the proportion of each deme using the sizes element
for calculation of median and quantiles (0.025 and 0.975).
}
\examples{
#TO DO
}
