% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_posterior.R
\name{mq_r0}
\alias{mq_r0}
\title{Calculates median and quantiles for basic reproduction number}
\usage{
mq_r0(data.gp, data.msm, times)
}
\arguments{
\item{data.gp}{matrix for basic reproduction number for general population}

\item{data.msm}{matrix for basic reproduction number for msm}

\item{times}{vector of time points}
}
\value{
dataframe of median, and upper and lower bounds for the posterior
}
\description{
Calculates median and quantiles for basic reproduction number
}
\details{
in both matrices (data.gp and data.msm), each column
   represents a different replicate and each row represents a time point.
}
\examples{
#TO DO
}
