% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/for_mcmc.R
\name{obj_fun}
\alias{obj_fun}
\title{Object function}
\usage{
obj_fun(parameters)
}
\arguments{
\item{parameters}{is a vector or matrix containing the proposals o f the mcmc}
}
\value{
the likelihood of the model based on the new parameter values
   (proposals of the mcmc)
}
\description{
This object function will receive the proposals of the mcmc.
The reason of using an object function is to make it easier to change the
values of the parameters to be estimated in THETA.
Note that not all parameters listed in THETA will be estimated
}
\details{
This object function uses a series of global variables that are declared
   in \code{\link{"analysis/scripts/model.R"}} and \code{\link{"analysis/scripts/load_data.R"}}
   These global variables are:
   \itemize{
   \item \code{THETA} is a list with parameter values
   \item \code{dated.tree} which is of class \code{\link[phydynR]{DatedTree}}
   \item \code{dm} the demographic model constructed with \code{\link[phydynR]{build.demographic.process}}
    \item \code{X0} vector that contains the initial conditions for the demes in our model
   }
}
\examples{
#In our example we would like to estimate 12 parameters. the parameter maleX will be fixed
# and the value for maleX is also changed in this obj_fun

lnl <- obj_fun(c(0.6, 0.4, 0.1, 1987, 0.4, 0.4, 0.2, 1995, 0.05, 0.1, 0.85, 0.85))

}
