% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dating.R
\name{organize_metadata_bySubtype}
\alias{organize_metadata_bySubtype}
\title{Organize metadata by HIV subtype.}
\usage{
organize_metadata_bySubtype(all_data_SN, all_data_CGR, code)
}
\arguments{
\item{all_data_SN}{Dataframe object of metadata for the Senegal only sequences}

\item{all_data_CGR}{Dataframe object of metadata for the CGR sequences}

\item{code}{If code = 1, it will return as a list dataframes for each subtype
but with dating formatted to decimal. If code = 2, it will return a list
of dataframes in which collection date is missing, and a dataframe with all
SN only available sequences}
}
\value{
a list of dataframes
}
\description{
This function will organize metadata for dating analysis purpose.
It will subset data for subtypes B, C and 02_AG, and have the collection
date formated as decimal for tip dating analysis
}
\examples{
TO DO
}
