% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_posterior.R
\name{posterior_trajectories}
\alias{posterior_trajectories}
\title{Get posterior distribution samples for the demographic model}
\usage{
posterior_trajectories(run, burnin, n, THETA)
}
\arguments{
\item{run}{object of class "mcmcSamplerList" and "bayesianOutput"}

\item{burnin}{intenger representing number of samples to remove as burnin}

\item{n}{integer representing the number of samples to get from the
posterior distribution}

\item{THETA}{list of parameter template used when running the MCMC analysis}
}
\value{
a list of solved objects for posterior distributions and MAP (maximum
   a posteriori)
}
\description{
Solve the demographic model for n combination of parameter values from the
posterior distribution
}
\examples{
# TO DO
}
\seealso{
\code{\link{post_traj_mx}}
}
