% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_posterior.R
\name{prop_infected}
\alias{prop_infected}
\title{Calculates the proportion of infected individuals}
\usage{
prop_infected(values, times)
}
\arguments{
\item{values}{dataframe of absolute numbers for median, lower and upper
bound for each deme}

\item{times}{vector of times points associated to the size list}
}
\value{
dataframe of the proportions of infected as median, lower and
   upper bounds
}
\description{
Calculates the proportion of infected individuals
}
\examples{
# TO DO
}
