% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence.R
\name{prPrevalenceStat}
\alias{prPrevalenceStat}
\title{Density function for the prevalence statistics}
\usage{
prPrevalenceStat(x)
}
\arguments{
\item{x}{is of class tfgy (after using the demographic function TO DO )}
}
\value{
density value for prevalence statistics
}
\description{
Calculates the density function for the proportion of infected men who are
msm
}
\examples{
#TO DO
}
