% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{read_mcmc_rds}
\alias{read_mcmc_rds}
\title{Read MCMC runs that is saved as RDS}
\usage{
read_mcmc_rds(fileNames)
}
\arguments{
\item{fileNames}{vector containg the file names}
}
\value{
as global variables the rds files assigned to a r object
}
\description{
Read the MCMC runs
}
\examples{
#TO DO
}
