% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{reorganize_deme_sizes}
\alias{reorganize_deme_sizes}
\title{Reorganize by deme size}
\usage{
reorganize_deme_sizes(Nrep, Ntime, sizes)
}
\arguments{
\item{Nrep}{integer for the number of replicates (number of combination of
parameter values based on the posterior distribution)}

\item{Ntime}{integer for the number of time points}

\item{sizes}{list with the sizes from the solved demographic model (TO DO)}
}
\value{
a list for each deme (gpm, gpf, and msm)
}
\description{
Reorganizes by deme based on the element size in the solved demographic model
}
\examples{
 #TO DO
}
