% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dating.R
\name{replace_values}
\alias{replace_values}
\title{Replace tip names used in a phylogentic tree}
\usage{
replace_values(metadata)
}
\arguments{
\item{metadata}{Object of class dataframe.}
}
\value{
Return HIV sample names by replacing to decimal format the part
   of the name separeted by a dash (_) which contains only the year of
   sample collection.
}
\description{
It replaces sample sequence names to reflect the input as expected
by PhyDyn in BEAST. This is an internal function used by other functions
in this package.
}
\examples{
To do
}
