% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dating.R
\name{root_drop_tips}
\alias{root_drop_tips}
\title{Re-root and remove tips in a phylogenetic tree}
\usage{
root_drop_tips(tree, outgroups)
}
\arguments{
\item{tree}{Object of class multiPhylo.}

\item{outgroups}{Vector of characters containing the names of the
outgroup sequences.}
}
\value{
a list of phylogenetic trees
}
\description{
Function that re-roots bootstrap trees and remove outgroup tips
(after re-rooting trees) in these trees.
}
\examples{
# TO DO
}
