% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence.R
\name{tfgy2prevalenceStat}
\alias{tfgy2prevalenceStat}
\title{Converts size for point estimate 2010 to prevalence of males that are msm}
\usage{
tfgy2prevalenceStat(tfgy)
}
\arguments{
\item{tfgy}{object of class tfgy. See demographic function (TO DO)}
}
\value{
numeric msm prevalence value
}
\description{
Converts size for point estimate 2010 to prevalence of males that are msm
}
\examples{
#To do
}
