% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{calculate_newCases}
\alias{calculate_newCases}
\title{Calculate the proportion of new cases}
\usage{
calculate_newCases(f)
}
\arguments{
\item{f}{list of object "births" as returned by the solved demographic model.
(see dm function in phydynR)}
}
\value{
the sum of columns in each matrix, which will correspond to new HIV
  cases. Here "births" object is a list in which each element corresponts to
  a time point and is a 4 by 4 matrix. See vignette for more details.
}
\description{
Calculate the proportion of new cases
}
\examples{
# TO DO
}
