% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dating.R
\name{data_format}
\alias{data_format}
\title{To format collection dates for samples}
\usage{
data_format(metadata, country)
}
\arguments{
\item{metadata}{Object of class dataframe. It will contain information
about collection dates for the sequences}

\item{country}{Object of class character. It should be in the form of
"SN" or "SN.CGR"}
}
\value{
A list in which the first element contain information in which the
    complete collection date is known; the second element of the list contain
    elements that have missing dating information
}
\description{
This function formats the dates in the form of YYYY-MM-DD to a decimal year to
be used in tip dating analysis.
}
\details{
The returned elements in the list will contain 2 dataframes. The
    first dataframe will have NA for collection dates that is only known
    from some parts (I mean, it is only know the YYYY.MM, or only YYYY). The
    second dataframe will have the upper and lower bound for the samples
    that we only partally know the collection dates.
}
\examples{
# write example here
}
