% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dating.R
\name{organize_metadata_bySubtype}
\alias{organize_metadata_bySubtype}
\title{Organize metadata by HIV subtype.}
\usage{
organize_metadata_bySubtype(all_data_SN, all_data_CGR, code)
}
\arguments{
\item{all_data_SN}{Object of class dataframe. This dataframe contains
metadata for Senegal sequences}

\item{all_data_CGR}{Object of class dataframe. This dataframe contains
metadata for CGR sequences}

\item{code}{If code = 1, it will return as a list dataframes for each subtype,
and collection dates are formatted to decimal. If code = 2, it will return
a list of dataframes for (1) all Senegal and CGR sequences and (2) samples
in which collection date is missing.}
}
\value{
a list of dataframes
}
\description{
Function that organizes metadata for dating analysis purpose.
It will subset data for subtypes B, C and CRF 02_AG, and collection
date will be formated as decimal for tip dating analysis
}
\examples{
TO DO
}
