% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{post_traj_mx}
\alias{post_traj_mx}
\title{Calculate the trajectories using posterior distrubution when also estimating
 maleX}
\usage{
post_traj_mx(parameters, THETA)
}
\arguments{
\item{parameters}{list of parameter estimation from the posterior distribution}

\item{THETA}{list of parameter values from the Model. Here the functions
for the spline function will also de loaded}
}
\value{
the solved object #to do
}
\description{
Solves the demographic model for each combination of parameter values
}
\examples{
#TO DO
}
