% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_posterior.R
\name{read_mcmc_rds}
\alias{read_mcmc_rds}
\title{Reads MCMC runs}
\usage{
read_mcmc_rds(fileNames)
}
\arguments{
\item{fileNames}{vector containg the file names}
}
\value{
global variables for each MCMC run as an R object
}
\description{
Read Markov chain Monte Carlo (MCMC) runs saved as RDS files
}
\examples{
#TO DO
}
