% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dating.R
\name{replace_values}
\alias{replace_values}
\title{Replace values in sequence tip name}
\usage{
replace_values(metadata)
}
\arguments{
\item{metadata}{Object of class dataframe.}
}
\value{
Return the sample names by replacing the last bit separeted by a
   dash (_) which represents the year of sample collection to the decimal
   representing the collection date.
}
\description{
It replaces values in sample sequence name to reflect the input as expected
by BEAST.
}
\examples{
To do
}
