% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence.R
\name{set_prev_state_parameters}
\alias{set_prev_state_parameters}
\title{Set states for prevalence values}
\usage{
set_prev_state_parameters()
}
\value{
mean (MEAN_PREV_STAT) and standard deviation (SD_PREV_STAT)
for prevalence
}
\description{
Function that sets the states for the prevalence values to calculate mean
and standard deviation to be used in MCMC runs
q = proportion of males who are msm
p_msm = msm HIV prevalence
p_m = heterosexual male (gpm) prevalence
X = proportion of infected gpm who are msm (assumes that the estimation
of p_msm is independent of estimated prevalence )
}
\examples{
#TO DO
}
